+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `_.sdk.inOrder` function to help specify the order in which your objects should be applied
 - `kct.io` external variable for the current environment, previously global `_`
 - `kct.libsonnet` to export things from `kct.io` more easily along with a future std